ID-243 refactored to extract test classes into their own modules
